pyproject.toml
[tool.poetry]
name = "novapipe"
version = "0.1.0"
description = "NovaPipe: An advanced, plugin-driven ETL CLI tool"
authors = ["Muqtar Shaikh <muqtar.shaikh225@gmail.com>"]
readme = "README.md"
packages = [{ include = "novapipe", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = "^8.1.0"
PyYAML = "^6.0"
pydantic = "^2.0"
jinja2 = "^3.1"     # add Jinja2 for templating
prometheus_client = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^5.0"
black = "^24.0"
mypy = "^1.0"

[tool.poetry.scripts]
novapipe = "novapipe.cli:cli"

[tool.poetry.group.dev.dependencies]
poetry-dynamic-versioning = "^1.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
